Step 1: Create the necesary files.
Step 1.1: Create the project, install the dependencies nodemon, express, cors.

Step 1.2: Update the packaje.json file to be type module.

Step 1.3: Create the index.js file, import all the modules.

Step 1.4: Declare the task array to store the tasks. Create the request for the server which is GET, POST , DELETE.

Step 1.5: Declare the port and assign it to the server.